7 lab

+++ app/Http/Controllers/ArticleController.php
@@ -25,6 +25,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', [self::class]);
         return view('articles/create');
     }
 
@@ -60,6 +61,7 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
+        $article->load(['comments.user']);
         return view('articles/showOne', ['article' => $article]);
     }
 
@@ -71,6 +73,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize('create', [self::class]);
         return view('articles/edit', ['article' => $article]);
     }
 
@@ -106,6 +109,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize('create', [self::class]);
         $article->delete();
         return redirect()->route('article.index');
     }
